objective3_interface/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.1.0
python-multipart==0.0.6
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
pydantic==2.5.0
PyJWT==2.8.0
email-validator==2.1.0
